Add test to the name of credits and fix travis problems
